delete MyTeamProfile component and don't solve MyTeamMemberPage -> MyTeamsProfilePage to props
